Changed all to integers

+++ src/algebra/Factor.java
@@ -7,10 +7,10 @@ import java.util.*;
  * @author UndeadScythes
  */
 public class Factor {
-    public static List<Long> getPrimeFactors(final long number) {
-        long test = number;
-        final List<Long> factors = new ArrayList<Long>();
-        for(long i = 2; i <= test / i; i++) {
+    public static List<Integer> getPrimeFactors(final int number) {
+        int test = number;
+        final List<Integer> factors = new ArrayList<Integer>();
+        for(int i = 2; i <= test / i; i++) {
             while(test % i == 0) {
                 factors.add(i);
                 test /= i;
@@ -22,13 +22,13 @@ public class Factor {
         return factors;
     }
 
-    public static List<Long> getFactors(final long number) {
-        final List<Long> factors = new ArrayList<Long>();
+    public static List<Integer> getFactors(final int number) {
+        final List<Integer> factors = new ArrayList<Integer>();
         int inc = 1;
         if(number % 2 != 0) {
             inc = 2;
         }
-        for(long i = 1; i <= number / 2; i += inc) {
+        for(int i = 1; i <= number / 2; i += inc) {
             if(number % i == 0) {
                 factors.add(i);
             }
